Implementando endpoint de estatisticas

+++ src/main/java/br/com/fiap/flux/video/domain/Estatistica.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Estatistica {
 
-    private Long quantidadeVideos;
+    private Integer quantidadeVideos;
     private Long quantidadeVideosFavoritos;
     private Long mediaVisualizacoes;
